# account-service/src/main/java/com/example/account_service/application/dto/CreateMovementRequest.java
package com.example.account_service.application.dto;

import java.math.BigDecimal;
import java.util.UUID;

public record CreateMovementRequest(
        UUID accountId,
        String type,
        BigDecimal value
) {}